requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
branca==0.8.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
et-xmlfile==1.1.0
folium==0.18.0
fonttools==4.48.1
frozenlist==1.7.0
gitdb==4.0.11
GitPython==3.1.41
gotrue==2.12.0
groq==0.4.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.1.0
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.1.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
multidict==6.5.0
numpy==1.26.3
openai==1.9.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
Pillow==9.5.0
plotly==5.18.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.10.1
Pympler==1.0.1
pyparsing==3.1.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
realtime==2.4.3
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
storage3==0.11.3
streamlit==1.46.0
streamlit-discourse==0.1.3
streamlit-option-menu==0.3.12
streamlit_folium==0.23.2
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
tenacity==8.2.3
termcolor==2.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing==3.7.4.3
typing_extensions==4.14.0
tzdata==2023.4
tzlocal==4.3.1
urllib3==2.1.0
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
websockets==14.2
xlrd==2.0.1
xyzservices==2024.9.0
yarl==1.20.1
zipp==3.17.0
